backend/requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.3
chromadb==0.6.3
Flask==3.1.0
Flask_Cors==5.0.0
Flask_Login==0.6.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
newspaper3k==0.2.8
numpy==2.2.4
openai==1.66.3
pandas==2.2.3
psycopg2==2.9.10
Requests==2.32.3
scikit_learn==1.6.1
selenium==4.29.0
SQLAlchemy==2.0.38
webdriver_manager==4.0.2
Werkzeug==3.1.3
